+++ frontend/my-adopthive-app/src/AdopterList/AdopterList.jsx
@@ -12,7 +12,6 @@ const AdopterList = () => {
   const { user } = useContext(UserContext);
   const [matchResult, setMatchResult] = useState(null);
   const [isSideBarOpen, setIsSideBarOpen] = useState(false);
-  const [favourites, setFavourites] = useState([]);
   const { isLoading, startLoading, stopLoading } = useLoading();
   const UserId = user.id
 
@@ -54,39 +53,10 @@ const AdopterList = () => {
         credentials: 'include',
         body: JSON.stringify({ preferences: preferenceList, UserId })
       });
-
       if (response.ok) {
         alert('Preferences saved successfully');
       } else {
         setError('Failed to save preferences');
-      }
-    } catch (error) {
-      setError('Error: ' + error.message);
-    }
-  };
-
-  const handleLike = async (profileId, isAdopter) => {
-    try {
-      if (favourites.includes(profileId)) {
-        const response = await fetch(`http://localhost:3004/favourites/${UserId}/${profileId}`, {
-          method: 'DELETE',
-        });
-        if (response.ok) {
-          setFavourites(favourites.filter(id => id !== profileId));
-        } else {
-          setError('Failed to remove from favourites');
-        }
-      } else {
-        const response = await fetch('http://localhost:3004/favourites', {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify({ UserId, adopterId: isAdopter ? profileId : null, adopteeId: !isAdopter ? profileId : null })
-        });
-        if (response.ok) {
-          setFavourites([...favourites, profileId]);
-        } else {
-          setError('Failed to add to favourites');
-        }
       }
     } catch (error) {
       setError('Error: ' + error.message);
@@ -143,14 +113,6 @@ const AdopterList = () => {
                       required
                     />
                   </div>
-                  <i
-                    className={`fa-solid fa-heart ${favourites.includes(adopter.id) ? 'liked' : ''}`}
-                    onClick={() => handleLike(adopter.id, true)}
-                    style={{
-                      color: favourites.includes(adopter.id) ? 'red' : 'gray',
-                      cursor: 'pointer'
-                    }}
-                  />
                 </li>
               ))}
             </ul>
