+++ frontend/my-adopthive-app/src/MainPage/MainPage.css
@@ -6,7 +6,7 @@
     height: 100vh;
     text-align: center;
     position: relative;
-    background-image: url(https://images.pexels.com/photos/9873934/pexels-photo-9873934.jpeg);
+    background-image: url(https://images.pexels.com/photos/8799247/pexels-photo-8799247.jpeg);
     background-size: cover;
     background-repeat: no-repeat;
 }
@@ -14,13 +14,13 @@
 .main-page h1 {
     font-size: 80px;
     margin-bottom: 1rem;
-    color: white;
+    color: black;
 }
 
 .main-page p {
     font-size: 25px;
     margin-bottom: 2rem;
-    color: white;
+    color: black;
 }
 
 .get-started-container {
@@ -31,8 +31,8 @@
 .get-started-btn {
     padding: 1rem 2rem;
     font-size: 1.2rem;
-    background-color: rgb(58, 21, 21);
-    color: #fff;
+    background-color: aquamarine;
+    color: black;
     border: none;
     border-radius: 5px;
     bottom: 20px;
